requirements.txt
Django>=2.0,<2.1
django-debug-toolbar
django-environ
psycopg2-binary
